+++ src/features/recovery-zone/constants/riskPalette.ts
