+++ src/features/recovery-zone/index.ts
@@ -1,4 +1,12 @@
-export { default as MapSeoulRisk } from './ui/MapSeoulRisk';
-export { default as MainContent } from './ui/maincontent/MainContent';
-export { default as BottomSheet } from './ui/bottomsheet/BottomSheet';
-export { BackgroundCircle } from './ui/MapBackground';
+export { default as MapSeoulRisk } from './components/MapBackground/MapSeoulRisk';
+export { default as MainContent } from './containers/MainContent/MainContent';
+export { default as BottomSheet } from './containers/BottomSheet/BottomSheet';
+export { BackgroundCircle } from './components/MapBackground/MapBackground';
+
+// Re-export components
+export { default as Card } from './components/Card/Card';
+export { default as NoticeCard } from './components/Card/NoticeCard';
+export { default as Header } from './components/Header/Header';
+
+// Re-export types
+export * from './types';
