+++ src/pages/RecoveryPage.tsx
@@ -1,4 +1,4 @@
-import { Wrapper } from "@features/recovery-zone/types/BasicElement";
+import { Wrapper } from "@features/recovery-zone/components/UI/BasicElement";
 import {
   BottomSheet,
   MainContent,
@@ -6,22 +6,32 @@ import {
 } from "@features/recovery-zone";
 import * as S from "@widget/MapWithOverlay/MapWithOverlay.styles";
 
-const BackgroundCirclePos = [
+// Background circle positions configuration
+const BACKGROUND_CIRCLE_POSITIONS = [
   { top: 238, left: -127 },
   { top: 70, left: 238 },
   { top: 491, left: -67 },
-];
+] as const;
 
+/**
+ * Recovery Page - 복구 현황 및 위험 지역 정보를 표시하는 메인 페이지
+ */
 const RecoveryPage = () => {
   return (
     <S.Wrap>
       <Wrapper $gap={40} $ColumnDirection={true}>
-        <MainContent></MainContent>
-        {/* <BottomSheet></BottomSheet> */}
+        <MainContent />
+        {/* TODO: BottomSheet 활성화 시 주석 해제 */}
+        {/* <BottomSheet /> */}
       </Wrapper>
 
-      {BackgroundCirclePos.map((pos, idx) => (
-        <BackgroundCircle key={idx} $top={pos.top} $left={pos.left} />
+      {/* Background decoration circles */}
+      {BACKGROUND_CIRCLE_POSITIONS.map((position, index) => (
+        <BackgroundCircle 
+          key={`bg-circle-${index}`} 
+          $top={position.top} 
+          $left={position.left} 
+        />
       ))}
     </S.Wrap>
   );
